chore(option): add serialization class and update associated tests
The toStringTag property is added for 'Option' classes to enable
serialization and proper string representation of object instances.
The changes also include an update of the 'None' and 'Some' classes
from implements to extends 'Option'. The corresponding unit tests
have been added to ensure the accurate operation of the new features.