fix(option): Refactor Option.None and fix related tests
The None method in the Option class was refactored, changing it to
allow for a type parameter. This resulted in a series of changes
across the class to accommodate the new function signature. Test
cases in the option.test.ts file were also updated to reflect this
change, leading to their passing statuses.