refactor(option): Refine `None` type annotations
This commit fine-tunes the type annotations in the option test and
the option module. It modifies the `expectTypeOf<Option.Value<typeof
noneOfUnknown>>().toEqualTypeOf<unknown>()` in the test to ensure it
equals 'never' instead of 'unknown'. It also sets a default type to
'never' for the None() constructor in the option module. Lastly, it
makes sure the test accurately reflects these type annotation changes.